requirements.txt
fastapi
uvicorn[standard]
pvlib
shapely
numpy
pandas
pyproj
python-dotenv
httpx
asyncpg
